englishbot トップページgrid化（レスポンシブ対応）
